rft: move data processing into run time
When running in multiprocessing mode, python will fork or spawn the
current process into sub process, essentially copying all the objects
already declared in the __init__ function. If coroutines are created
inside __init__, the original copy of those coroutines will not be run,
and result in a warning of 'coroutine is never awaited'. To support
light weight multiprocessing mode, move evrything to start/run method if
possible, leaving only essentials in declaration.